Move test to `mod.rs`
